In the end cet::search_path wasn't up to the task of finding glob'ed (i.e. wildcarded) flux files in the way we wanted and was compatible with GENIE.  Introduced new function FindFluxPath() to do it.
Using the the FW_SEARCH_PATH list of directories, apply the
user supplied pattern as a suffix to find the flux files.
The userpattern might include simple wildcard globs (in contrast
to proper regexp patterns) for the filename part, but not any
part of the directory path.  If files are found in more than
one FW_SEARCH_PATH alternative take the one with the most files.


git-svn-id: http://cdcvs.fnal.gov/subversion/nusoftsvn/trunk/nutools@123 692d5501-4a7f-4c20-9605-99b1da982744